requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
python-dotenv==1.0.1
email-validator==2.2.0
